Finished the draft API & some logic and style changes on front end.

+++ src/main/kotlin/org/kinleoapple/plugins/database/relation/Blog.kt
@@ -3,16 +3,13 @@ package org.kinleoapple.plugins.database.relation
 import io.ktor.server.application.*
 import org.kinleoapple.plugins.database.Database
 import org.kinleoapple.plugins.database.configureDatabase
-import org.ktorm.schema.Table
-import org.ktorm.schema.datetime
-import org.ktorm.schema.int
-import org.ktorm.schema.varchar
+import org.ktorm.schema.*
 
 private const val TABLE_NAME = "blog"
 
 data object Blog: Table<Nothing>(TABLE_NAME) {
-    val blogId = int("blog_id").primaryKey()
-    val catId = int("cat_id")
+    val blogId = long("blog_id").primaryKey()
+    val catId = long("cat_id")
     val blogPubDt = datetime("blog_pub_dt")
     val blogPath = varchar("blog_path")
     val blogTitle = varchar("blog_title")
@@ -21,8 +18,8 @@ data object Blog: Table<Nothing>(TABLE_NAME) {
 
 private val sql: String = """
     CREATE TABLE IF NOT EXISTS `${TABLE_NAME}` (
-          `blog_id` INT NOT NULL,
-          `cat_id` INT NOT NULL,
+          `blog_id` BIGINT NOT NULL,
+          `cat_id` BIGINT NOT NULL,
           `blog_pub_dt` DATETIME NOT NULL,
           `blog_path` VARCHAR(100) NOT NULL,
           `blog_title` VARCHAR(100) NOT NULL,
