+++ src/main/kotlin/org/kinleoapple/plugins/database/relation/User.kt
@@ -9,21 +9,22 @@ import org.ktorm.dsl.deleteAll
 import org.ktorm.dsl.insert
 import org.ktorm.schema.Table
 import org.ktorm.schema.int
+import org.ktorm.schema.long
 import org.ktorm.schema.varchar
 
 private const val TABLE_NAME = "user"
 
 data object User : Table<Nothing>(TABLE_NAME) {
-    val userId = int("user_id").primaryKey()
-    val quoteId = int("quote_id")
+    val userId = long("user_id").primaryKey()
+    val quoteId = long("quote_id")
     val userName = varchar("user_name")
     val userPassword = varchar("user_password")
 }
 
 private val sql: String = """
     CREATE TABLE IF NOT EXISTS `${TABLE_NAME}` (
-        `user_id` INT NOT NULL DEFAULT 0, 
-        `quote_id` INT NOT NULL DEFAULT 0,
+        `user_id` BIGINT NOT NULL DEFAULT 0, 
+        `quote_id` BIGINT NOT NULL DEFAULT 0,
         `user_name` VARCHAR(30) NOT NULL,
         `user_password` VARCHAR(40) NOT NULL,
     PRIMARY KEY (`user_id`)
